requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
requests>=2.31.0
google-analytics-data>=0.18.0
gspread>=5.12.0
oauth2client>=4.1.3
pandas>=2.2.2
python-dateutil>=2.8.2
